remove old files from previous versions
previous versions of xivo-dist and wazo-dist did not use dh_python properly and
the prerm did not get generated properly to remove the pyc files.

That left the old package directories on the file system. The old packages
cannot be repaired but we can still clean up the mess behind.